pyproject.toml
[project]
name = "grok-mcp-server"
version = "1.0.0"
description = "A production-ready MCP server for Grok AI models"
authors = [{name = "Igor Warzocha", email = "igor.warzocha@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["mcp", "grok", "ai", "llm", "fastmcp"]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "pillow>=11.0.0",
    "numpy>=2.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0"
]

[project.scripts]
grok-mcp = "src.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "YTT", "B", "A", "COM", "C4", "DTZ", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "ERA", "PGH", "PLE", "PLW", "RUF"]
ignore = ["E501", "B008", "B905"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]